chore: Refactor app routes
