+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export * from './AddFriendForm';
+export * from './Button';
+export * from './Friend';
+export * from './FriendList';
+export * from './SplitBillForm';
+export * from './App';
