+++ src/global.css
